refactor: domain package 분리
