+++ src/main/java/study/step3/domain/RealCar.java
@@ -1,4 +1,4 @@
-package study.step3;
+package study.step3.domain;
 
 import java.util.Objects;
 import java.util.Random;
