+++ src/test/java/study/step3/ResultViewTest.java
@@ -8,6 +8,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.ArgumentsProvider;
 import org.junit.jupiter.params.provider.ArgumentsSource;
+import study.step3.domain.RaceRecord;
 
 import java.util.*;
 import java.util.stream.Stream;
